Fix warning about unused mut
For field-less structs and enums, the mutability of `bytes` in `exhume`
is warned about. This is inconvenient for users, as they have to
annotate the struct with an `#[allow(unused_mut)` lint. The issue only
occurred using the nightly Rust compiler.

This change simply annotates the exhume method like that, so that the
warning is no longer shown to users. To prevent regressions `unused_mut`
is now a hard error in the test module.