Upgrade to Rust 2018 and bump dependencies accordingly
The crate was using a sufficiently old version of syn that the macro
would panic if it was used on a struct that contained a `crate::` path.
Upgrading to the latest version of syn fixes the issue; see the enclosed
test.

For good measure, this commit also upgrades to the latest version of
quote, which in turn requires introducing a dependency on proc_macro2,
and synstructure, and bumps the edition of the crate to Rust 2018.

This is part of a push to upgrade Timely and Differential to Rust 2018.