Upgrade proc-macro2, quote, and synstructure dependencies
This patch updates the proc-macro2, quote, and synstructure dependencies
to more recent versions. The former two crates, in particular, have
reached 1.0 versions which are used by an increasing amount of other
crates in the ecosystem and so going with the flow can reduce the number
of crates effectively pulled in during compilation.
The dependency to syn is removed completely, as it is not used.